feat: update auth middleware
