+++ src/lib/middleware/index.ts
@@ -1,2 +1,2 @@
+export * from './auth';
 export * from './error';
-export * from './persist-login';
